Aggiunta una view che permetta all'admin di visualizzare gli ordini fatti da parte dei clienti. Aggiunto Logout per tornare alla view del login. Aggiunta una label per mostrare il budget (guadagni-investimenti) dell'admin. Aggiunto username ed immagine per il customer dopo aver effettuato il login. Aggiunti vari messaggi di conferma e messaggi di verifica del checkout

+++ src/main/java/com/restaurant/administrator/StaffController.java
@@ -7,11 +7,8 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.control.Label;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
+import javafx.scene.control.*;
 import javafx.fxml.FXML;
-import javafx.scene.control.TextField;
 import javafx.stage.Stage;
 
 import java.sql.SQLException;
@@ -29,6 +26,7 @@ public class StaffController {
     @FXML private TableColumn<Staff, String> colStaffRole;
     @FXML private TableColumn<Staff, String> colStaffSalary;
     @FXML private TableColumn<Staff, String> colStaffDate;
+    @FXML private Button returnButton;
 
     @FXML
     public void initialize() throws Exception{
@@ -77,10 +75,10 @@ public class StaffController {
             if(staff != null) {
                 StaffUpdateController.setCurrent(staff);
                 Parent root = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("StaffUpdate.fxml")));
-                Scene admStaff = new Scene(root);
+
                 //get stage information
                 Stage window = new Stage();
-                window.setScene(admStaff);
+                window.setScene(new Scene(root));
                 window.show();
             }
         }
@@ -95,9 +93,16 @@ public class StaffController {
         try{
             Staff staff = staffTable.getSelectionModel().getSelectedItem();
             if(staff != null){
+                Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+                alert.setTitle("Confirm Dismissal");
+                alert.setHeaderText("Do you really want to fire this employee?");
+                alert.setContentText("Press OK to fire the current employee");
+
+                if (alert.showAndWait().get() == ButtonType.OK) {
                 StaffDAO.deleteStaffByID(staff.getId());
                 ObservableList<Staff> staffList = StaffDAO.getAllRecords();
                 populateTable(staffList);
+                }
             }
         }
         catch (SQLException e){
@@ -124,8 +129,12 @@ public class StaffController {
             }
         }
     }
+
     public void return_back(ActionEvent event)  throws Exception{
         try {
+            Stage stage = (Stage) returnButton.getScene().getWindow(); // Ottieni il riferimento alla finestra di AdministratorDish
+            stage.close(); // Chiudi la finestra di AdministratorDish
+
             Parent root = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("AdministratorView.fxml")));
             Scene admStaff = new Scene(root);
